+++ example/lib/main.dart
@@ -1,20 +1,22 @@
 import 'package:example/screens/status_alert_screen.dart';
 import 'package:flutter/material.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Flutter Demo',
       theme: ThemeData(
-        primarySwatch: Colors.deepPurple,
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         brightness: Brightness.light,
       ),
       darkTheme: ThemeData(brightness: Brightness.dark),
-      home: StatusAlertScreen(),
+      home: const StatusAlertScreen(),
     );
   }
 }