+++ lib/status_alert.dart
@@ -1,5 +1,3 @@
-library status_alert;
-
 export 'src/models/status_alert_media_configuration.dart';
 export 'src/models/status_alert_text_configuration.dart';
 export 'src/status_alert.dart';